move attachment into attachments module
